% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.RasterLayer.R
\name{as.RasterBrick}
\alias{as.RasterBrick}
\title{as.RasterBrick}
\usage{
as.RasterBrick(x, res = NULL)
}
\arguments{
\item{x}{MAgPIE object}

\item{res}{spatial data resolution. If not provided it will be guessed.}
}
\value{
A RasterBrick object
}
\description{
Convert magclass object to a RasterBrick object
}
\author{
Jan Philipp Dietrich
}
