% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.RasterLayer.R
\name{as.RasterLayer}
\alias{as.RasterLayer}
\title{as.RasterLayyer}
\usage{
as.RasterLayer(x, res = NULL)
}
\arguments{
\item{x}{MAgPIE object}

\item{res}{spatial data resolution. If not provided it will be guessed.}
}
\value{
A RasterLayer object
}
\description{
Convert magclass object to a RasterLayer object
}
\author{
Jan Philipp Dietrich
}
