% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SpatRasterDataset.R
\name{as.SpatRasterDataset}
\alias{as.SpatRasterDataset}
\title{as.SpatRasterDataset}
\usage{
as.SpatRasterDataset(...)
}
\arguments{
\item{...}{arguments passed to as.SpatRaster}
}
\value{
A SpatRasterDataset object
}
\description{
Convert magclass object to a SpatRasterDataset object. Requires the terra package.
}
\details{
Calls \code{\link{as.SpatRaster}} and then \code{\link{spatRasterToDataset}}.
}
\seealso{
Other MAgPIE-Conversions:
\code{\link[=as.RasterBrick]{as.RasterBrick()}},
\code{\link[=as.SpatRaster]{as.SpatRaster()}},
\code{\link[=as.SpatVector]{as.SpatVector()}},
\code{\link{as.array-methods}},
\code{\link{as.data.frame-methods}},
\code{\link[=as_tibble.magpie]{as_tibble.magpie()}},
\code{\link[=unwrap]{unwrap()}},
\code{\link[=wrap]{wrap()}}
}
\author{
Pascal Sauer
}
\concept{MAgPIE-Conversions}
