% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.magpie.R
\name{cbind.magpie}
\alias{cbind.magpie}
\title{cbind method for MAgPIE objects}
\usage{
\method{cbind}{magpie}(..., deparse.level = 1)
}
\arguments{
\item{...}{see \link[base]{cbind}}

\item{deparse.level}{see \link[base]{cbind}}
}
\value{
A matrix as if cbind was applied to the internal vector of the MAgPIE objects
}
\description{
cbind method for MAgPIE-objects. Issues a warning that mbind should be used
if all objects are magpie objects and continues afterwards.
}
\seealso{
Other DataBinding:
\code{\link[=extend]{extend()}},
\code{\link[=magpie_expand]{magpie_expand()}},
\code{\link[=magpiesort]{magpiesort()}},
\code{\link[=matchDim]{matchDim()}},
\code{\link[=mbind]{mbind()}}
}
\author{
Patrick Rein
}
\concept{DataBinding}
