% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonYears.R
\name{commonYears}
\alias{commonYears}
\title{commonYears}
\usage{
commonYears(..., asInteger = FALSE)
}
\arguments{
\item{...}{One or more magclass objects, or alternatively a list of magclass objects.}

\item{asInteger}{Passed on to getYears. If TRUE the output format is 1995 otherwise "y1995".}
}
\value{
The years (as returned by getYears) the input objects have in common.
}
\description{
Given (a list of) magclass objects, return the years (getYears) they have in common,
ignoring magclass objects with 0 or 1 years.
}
\author{
Pascal Sauer
}
