% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_magpie.R
\name{complete_magpie}
\alias{complete_magpie}
\title{complete_magpie}
\usage{
complete_magpie(x, fill = NA, dim = 3)
}
\arguments{
\item{x}{MAgPIE object which should be completed.}

\item{fill}{Value that shall be written into the missing entries}

\item{dim}{dimensions in which the completion should take place (1, 2 and/or 3). For
full completion use \code{1:3}}
}
\value{
The completed MAgPIE object
}
\description{
MAgPIE objects can be incomplete to reduce memory. This function blows up a
magpie object to its real dimensions, so you can apply unwrap.
}
\examples{

pop <- maxample("pop")
complete_magpie(pop)

ani <- maxample("animal")
complete_magpie(ani)
}
\seealso{
\code{\link{add_dimension}},\code{\link{clean_magpie}}
}
\author{
Jan Philipp Dietrich, Benjamin Bodirsky
}
