% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{coord}
\usage{
coord(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
data frame with columns "lon" and "lat" containing
the coordinates of all spatial cells.
}
\description{
Extracts coordinates from a magpie object. Only works for objects
which provide coordinates as spatial dimensions with the names "lon"
and "lat"!
}
\details{
Please note that "." have to be replaced with "," in lon/lat 
coordinates as "." is reserved as divider between subdimensions.
}
\examples{
m <- new.magpie(c("0,5.0,5","0,5.-0,5","0,5.1,0","1,0.1,0"),sets=c("lon","lat","year","data"))
coord(m)
}
\seealso{
\code{\link{getItems}}
}
\author{
Jan Philipp Dietrich
}
