% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimExists.R
\name{dimExists}
\alias{dimExists}
\title{dimExists}
\usage{
dimExists(dim, x, sep = ".")
}
\arguments{
\item{dim}{A vector of dimension numbers or dimension names which should be
checked for.}

\item{x}{MAgPIE object in which the dimensions should be searched for.}

\item{sep}{A character separating joined dimension names.}
}
\value{
Boolean indicating whether the dimension exists or not.
}
\description{
This function checks whether a dimension given by name or number exists.
}
\examples{

pop <- maxample("pop")
dimExists(c("t", "scenario", "blablub"), pop)

}
\seealso{
\code{\link{dimCode}}

Other ObjectInfo:
\code{\link[=fulldim]{fulldim()}},
\code{\link[=hasCoords]{hasCoords()}},
\code{\link[=hasSets]{hasSets()}},
\code{\link[=is.temporal]{is.temporal()}},
\code{\link[=isYear]{isYear()}},
\code{\link[=ncells]{ncells()}},
\code{\link[=ndim]{ndim()}},
\code{\link[=sameDims]{sameDims()}}
}
\author{
Jan Philipp Dietrich
}
\concept{ObjectInfo}
