\name{.escapeRegex}
\alias{.escapeRegex}
\title{.escapeRegex}
\description{Escapes all symbols in a string which have a special meaning in regular expressions.}
\usage{.escapeRegex(x)}
\arguments{
  \item{x}{String or vector of strings that should be escaped.}
}
\value{
  The escaped strings.}
\author{Jan Philipp Dietrich}
\seealso{\code{\link[base]{grep}}}
