% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{extend}
\usage{
extend(
  x,
  xRange = c(-179.75, 179.75),
  yRange = c(89.75, -89.75),
  res = NULL,
  checkInRange = TRUE
)
}
\arguments{
\item{x}{A magpie object}

\item{xRange}{Range of x coordinates, in ascending order when writing to netCDF}

\item{yRange}{Range of y coordinates, in descending order when writing to netCDF}

\item{res}{Resolution of the data, if not provided it will be guessed}
}
\value{
An extended magpie object with the same data as x and gaps filled with NA
}
\description{
Extend a magpie object to a dense grid based on the given xRange, yRange and resolution.
This is e.g. required when writing netCDF files. Extending a sparse magpie object to a dense grid
requires much more memory, so use with caution.
}
\author{
Pascal Sauer
}
