% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulldim.R
\name{fulldim}
\alias{fulldim}
\title{Reconstructs full dimensionality of MAgPIE objects}
\usage{
fulldim(x, sep = ".")
}
\arguments{
\item{x}{A MAgPIE-object}

\item{sep}{A character separating joined dimension names}
}
\value{
A list containing in the first element the dim output and in the
second element the dimnames output of the reconstructed array.
}
\description{
Deprecated: Use \code{\link{getItems}} instead.
}
\details{
If a MAgPIE object is created from a source with more than one data
dimension, these data dimensions are combined to a single dimension. fulldim
reconstructs the original dimensionality and reports it.
}
\seealso{
\code{\link{as.magpie}},\code{\link{unwrap}},\code{\link{wrap}}

Other ObjectInfo:
\code{\link[=dimExists]{dimExists()}},
\code{\link[=hasCoords]{hasCoords()}},
\code{\link[=hasSets]{hasSets()}},
\code{\link[=is.temporal]{is.temporal()}},
\code{\link[=isYear]{isYear()}},
\code{\link[=ncells]{ncells()}},
\code{\link[=ndim]{ndim()}},
\code{\link[=sameDims]{sameDims()}}
}
\author{
Jan Philipp Dietrich
}
\concept{ObjectInfo}
