% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCells.R, R/setCells.R
\name{getCells}
\alias{getCells}
\alias{getCells<-}
\alias{setCells}
\title{Get Cells}
\usage{
getCells(x)

getCells(x) <- value

setCells(object, nm = "GLO")
}
\arguments{
\item{x, object}{MAgPIE object}

\item{value, nm}{cell names the data should be set to.}
}
\value{
getCells returns cell names of the MAgPIE-object, whereas setCells
returns the MAgPIE object with the manipulated cell names.
}
\description{
Extracts cell names of a MAgPIE-object
}
\details{
setCells is a shortcut to use a MAgPIE object with manipulated cell names.
setCells uses the variable names "object" and "nm" in order to be consistent
to the already existing function setNames.
}
\section{Functions}{
\itemize{
\item \code{getCells(x) <- value}: set cell names

\item \code{setCells()}: set cell names

}}
\examples{

a <- as.magpie(1)
 getCells(a)
 setCells(a, "AFR")
}
\seealso{
\code{\link{getRegions}}, \code{\link{getNames}},
\code{\link{setNames}}, \code{\link{getCPR}}, \code{\link{read.magpie}},
\code{\link{write.magpie}}, \code{"\linkS4class{magpie}"}

Other GetterSetter:
\code{\link[=getCPR]{getCPR()}},
\code{\link[=getComment]{getComment()}},
\code{\link[=getCoords]{getCoords()}},
\code{\link[=getDim]{getDim()}},
\code{\link[=getItems]{getItems()}},
\code{\link[=getNames]{getNames()}},
\code{\link[=getRegionList]{getRegionList()}},
\code{\link[=getRegions]{getRegions()}},
\code{\link[=getSets]{getSets()}},
\code{\link[=getYears]{getYears()}},
\code{\link[=setItems]{setItems()}},
\code{\link{setNames-methods}}
}
\author{
Jan Philipp Dietrich
}
\concept{GetterSetter}
