% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R, R/setComment.R
\name{getComment}
\alias{getComment}
\alias{getComment<-}
\alias{setComment}
\title{getComment}
\usage{
getComment(x)

getComment(x) <- value

setComment(object, nm = NULL)
}
\arguments{
\item{x, object}{MAgPIE object}

\item{value, nm}{A vector containing the comment.}
}
\value{
getComment returns the comment attached to a MAgPIE-object, NULL if
no comment is present. setComment returns the magpie object with the
modified comment.
}
\description{
Extracts the comment from a MAgPIE-object
}
\section{Functions}{
\itemize{
\item \code{getComment(x) <- value}: set comment

\item \code{setComment()}: set comment

}}
\examples{

a <- as.magpie(1)
 # returns NULL
 getComment(a)
 # set the comment
 getComment(a) <- c("bla", "blubb")
 getComment(a)

}
\seealso{
\code{\link{getRegions}}, \code{\link{getNames}},
\code{\link{getYears}}, \code{\link{getCPR}}, \code{\link{read.magpie}},
\code{\link{write.magpie}}, \code{"\linkS4class{magpie}"}

Other GetterSetter:
\code{\link[=getCPR]{getCPR()}},
\code{\link[=getCells]{getCells()}},
\code{\link[=getCoords]{getCoords()}},
\code{\link[=getDim]{getDim()}},
\code{\link[=getItems]{getItems()}},
\code{\link[=getNames]{getNames()}},
\code{\link[=getRegionList]{getRegionList()}},
\code{\link[=getRegions]{getRegions()}},
\code{\link[=getSets]{getSets()}},
\code{\link[=getYears]{getYears()}},
\code{\link[=setItems]{setItems()}},
\code{\link{setNames-methods}}
}
\author{
Markus Bonsch
}
\concept{GetterSetter}
