% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\alias{getCoords<-}
\title{Get Coordinates}
\usage{
getCoords(x, xlab = "x", ylab = "y")

getCoords(x, xlab = "x", ylab = "y") <- value
}
\arguments{
\item{x}{MAgPIE object}

\item{xlab}{label of x-dimension}

\item{ylab}{label of y-dimension}

\item{value}{coordinates as two column data.frame the data should be set to
(first column = x, second column = y).}
}
\value{
coordinates of the MAgPIE-object
}
\description{
Extracts spatial coordinates of a MAgPIE-object
}
\section{Functions}{
\itemize{
\item \code{getCoords(x, xlab = "x", ylab = "y") <- value}: set coordinates

}}
\examples{
a <- maxample("animal")
getCoords(a)
}
\seealso{
\code{\link{as.RasterBrick}}, \code{\link{getItems}}, \code{"\linkS4class{magpie}"}

Other GetterSetter:
\code{\link[=getCPR]{getCPR()}},
\code{\link[=getCells]{getCells()}},
\code{\link[=getComment]{getComment()}},
\code{\link[=getDim]{getDim()}},
\code{\link[=getItems]{getItems()}},
\code{\link[=getNames]{getNames()}},
\code{\link[=getRegionList]{getRegionList()}},
\code{\link[=getRegions]{getRegions()}},
\code{\link[=getSets]{getSets()}},
\code{\link[=getYears]{getYears()}},
\code{\link[=setItems]{setItems()}},
\code{\link{setNames-methods}}
}
\author{
Jan Philipp Dietrich
}
\concept{GetterSetter}
