% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPR.R
\name{getCPR}
\alias{getCPR}
\title{Get cells per region}
\usage{
getCPR(x)
}
\arguments{
\item{x}{MAgPIE object or a resolution written as numeric (currently only
data for 0.5 degree resolution is available).}
}
\value{
cells per region
}
\description{
Counts how many cells each region has and returns it as vector
}
\examples{

# a <- read.magpie("example.mz")
# getCPR(a)
getCPR(0.5)

}
\seealso{
\code{\link{getRegions}}, \code{\link{read.magpie}},
\code{\link{write.magpie}}
}
\author{
Jan Philipp Dietrich
}
