% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItems.R
\name{getItems}
\alias{getItems}
\title{Get Items}
\usage{
getItems(x, dim)
}
\arguments{
\item{x}{MAgPIE object}

\item{dim}{Dimension for which the items should be returned. Either number or 
name of dimension. See \code{\link{dimCode}} for more details.}
}
\value{
items of the requested dimension in the MAgPIE-object
}
\description{
Extract items of a given (sub-)dimension of a MAgPIE-object
}
\examples{
getItems(population_magpie,"scenario")
getItems(population_magpie,3.1)
}
\seealso{
\code{\link{dimCode}}
}
\author{
Jan Philipp Dietrich
}
