% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionList.R
\name{getRegionList}
\alias{getRegionList}
\alias{getRegionList<-}
\title{Get a list of celluare region-belongings}
\usage{
getRegionList(x)

getRegionList(x) <- value
}
\arguments{
\item{x}{MAgPIE object}

\item{value}{A vector with ncell elements containing the regions of each
cell.}
}
\value{
A vector with ncell elements containing the region of each cell.
}
\description{
Extracts a vector containing the region of each cell of a MAgPIE-object
}
\section{Functions}{
\itemize{
\item \code{getRegionList(x) <- value}: set region names

}}
\examples{

# a <- read.magpie("example.mz")
# getRegionList(a)
}
\seealso{
\code{\link{getRegions}},\code{\link{getYears}},
\code{\link{getNames}}, \code{\link{getCPR}}, \code{\link{read.magpie}},
\code{\link{write.magpie}}, \code{"\linkS4class{magpie}"}

Other GetterSetter:
\code{\link[=getCPR]{getCPR()}},
\code{\link[=getCells]{getCells()}},
\code{\link[=getComment]{getComment()}},
\code{\link[=getCoords]{getCoords()}},
\code{\link[=getDim]{getDim()}},
\code{\link[=getItems]{getItems()}},
\code{\link[=getNames]{getNames()}},
\code{\link[=getRegions]{getRegions()}},
\code{\link[=getSets]{getSets()}},
\code{\link[=getYears]{getYears()}},
\code{\link[=setItems]{setItems()}},
\code{\link{setNames-methods}}
}
\author{
Jan Philipp Dietrich
}
\concept{GetterSetter}
