% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessResolution.R
\name{guessResolution}
\alias{guessResolution}
\title{guessResolution}
\usage{
guessResolution(x)
}
\arguments{
\item{x}{A magpie object or the coordinates of a magpie object (the result
of \code{\link{getCoords}})}
}
\value{
The guessed resolution of the data
}
\description{
Guess the resolution of the given magpie object/coordinates by looking at the minimum
difference between unique sorted values. Fall back to 0.5 if guess is infinite.
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
