% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasSets.R
\name{hasSets}
\alias{hasSets}
\title{Has Sets}
\usage{
hasSets(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
Boolean indicating whether coordinates were found or not
}
\description{
Checks, whether set names have been set
}
\examples{

hasSets(maxample("pop"))
hasSets(maxample("animal"))

}
\seealso{
\code{\link{getCoords}}

Other ObjectInfo:
\code{\link[=dimExists]{dimExists()}},
\code{\link[=fulldim]{fulldim()}},
\code{\link[=hasCoords]{hasCoords()}},
\code{\link[=is.temporal]{is.temporal()}},
\code{\link[=isYear]{isYear()}},
\code{\link[=ncells]{ncells()}},
\code{\link[=ndim]{ndim()}},
\code{\link[=sameDims]{sameDims()}}
}
\author{
Jan Philipp Dietrich
}
\concept{ObjectInfo}
