% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unit_installed.R
\name{is_unit_installed}
\alias{is_unit_installed}
\title{is_unit_installed (!experimental!)}
\usage{
is_unit_installed(char)
}
\arguments{
\item{char}{A character string to be checked for units compatibility}
}
\value{
Returns a boolean. TRUE if char is recognized by the units package and FALSE otherwise.
If FALSE, char can be installed as a compatible unit via install_magpie_units.
}
\description{
This function quickly checks whether a character is already recognizable as a units object.
If FALSE, the unit can be installed via install_magpie_units.
}
\author{
Stephen Bi
}
