% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-method.R
\docType{methods}
\name{log-methods}
\alias{log-methods}
\alias{log,magpie-method}
\alias{logb-methods}
\alias{logb,magpie-method}
\title{log-method for MAgPIE objects}
\usage{
\S4method{log}{magpie}(x, base = exp(1))
}
\arguments{
\item{x}{a magpie object}

\item{base}{ia positive or complex number: the base with respect to which
logarithms are computed. Defaults to e=exp(1).}
}
\description{
log-method for MAgPIE-objects respectively. Works exactly as for arrays.
}
\section{Methods}{
 \describe{

\item{x = "magpie"}{ works as log(x) for arrays. }

}
}

