% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\name{magclassdata}
\alias{magclassdata}
\title{magclassdata}
\description{
General magclass-dataset
}
\details{
Please do not directly access that data. It should be only used by library
functions.
}
\author{
Jan Philipp Dietrich
}
