% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax-method.R, R/pmin-method.R
\docType{methods}
\name{pmax,magpie-method}
\alias{pmax,magpie-method}
\alias{magpie-pmin-pmax}
\alias{pmin}
\alias{pmax}
\alias{pmin,magpie-method}
\alias{pmin,ANY-method}
\alias{pmax,ANY-method}
\title{pmin/pmax}
\usage{
\S4method{pmax}{magpie}(..., na.rm = FALSE)

\S4method{pmin}{magpie}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Multiple magpie objects with dimensions that each are either the same or only
contain one item among all objects.}

\item{na.rm}{Passed to \code{\link[base]{pmin}}}
}
\value{
A new magpie object that contains the minimum/maximum of values between all the passed magpie objects.
}
\description{
magclass-aware pmin/pmax, which calculate the parallel minima/maxima of the input values
}
\details{
\code{\link[base]{pmin}}/\code{\link[base]{pmax}} do not care about the order of items in a magclass
object's dim. magclass pmin/pmax reorder items in each dim so they are in the same order.
}
\author{
Pascal Sauer, Patrick Rein
}
\keyword{methods}
