% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpieComp.R
\name{magpieComp}
\alias{magpieComp}
\title{magpieComp}
\usage{
magpieComp(bench, comp, reg = NA)
}
\arguments{
\item{bench}{A \code{MAgPIE} object.}

\item{comp}{A \code{MAgPIE} object.}

\item{reg}{The region(s) you want to focus on}
}
\value{
a list containing a1) the names found only in bench, a2) the names
found only in comp, b) a sorted data frame with the largest relative
difference between bench and comp in percentage values, and c) a magclass
object with the same values
}
\description{
Function that compares two magpie objects.
}
\details{
Function that compares two magpie objects.
}
\author{
Anastasis Giannousakis
}
