% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDim.R
\name{matchDim}
\alias{matchDim}
\title{match object dimensions of a magclass object to the dimensions of a reference object}
\usage{
matchDim(x, ref, dim = NULL, fill = NA)
}
\arguments{
\item{x}{a magclass object to be modified}

\item{ref}{a magclass object used as a reference for the modification}

\item{dim}{subset of dimensions for which the matching should be done. Either number or
name of dimension or a vector of these. See \code{\link{dimCode}} for more details.
Set to NULL (default) for matching all dimensions.}

\item{fill}{value to be set in new dimensions}
}
\description{
A helper that restricts and expands a magclass object x to the size of a magclass object ref.
Dimension names not present in x are added and set to the value provided by 'fill'.
Dimension names not present in ref are cropped.
}
\author{
Falk Benke
}
