% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmax.R, R/mpmin.R
\name{mpmax}
\alias{mpmax}
\alias{mpmin}
\title{mpmin/mpmax}
\usage{
mpmax(...)

mpmin(...)
}
\arguments{
\item{...}{Multiple magpie objects with the same dimensions.}
}
\value{
A new magpie object that contains the minimum/maximum of values between all the passed magpie objects.
}
\description{
magclass-aware pmin/pmax, which calculate the parallel minima/maxima of the input values
}
\details{
\code{\link{pmin}}/\code{\link{pmax}} do not care about the order of items in a magclass
object's dim. mpmin/mpmax reoder items in each dim so they are in the same order.
}
\author{
Pascal Sauer, Patrick Rein
}
