% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dim_convention.R
\name{old_dim_convention}
\alias{old_dim_convention}
\title{old_dim_convention}
\usage{
old_dim_convention(dim)
}
\arguments{
\item{dim}{The dim number in the new convention}
}
\value{
The dim number according to the old convention
}
\description{
Transforms new dim convention (e.g. 3.2) into old dim convention(e.g. 4)
}
\examples{

  dim=old_dim_convention(3.2)
  dim=old_dim_convention(1.1)

}
\seealso{
\code{\link{add_columns}},\code{\link{add_dimension}}
}
\author{
Benjamin Bodirsky
}
