% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_x_in_y.R
\name{place_x_in_y}
\alias{place_x_in_y}
\title{place_x_in_y}
\usage{
place_x_in_y(x, y, expand = T)
}
\arguments{
\item{x}{Object to be placed.}

\item{y}{Object in which x shall be placed}

\item{expand}{T: if x is larger than y, new columns are added.}
}
\value{
The combination of x and y. x overwrites y values which are in the
same place.
}
\description{
Function positions magpie object x into magpie object y.
}
\examples{

 pop <- maxample("pop")
 x <- pop[,"y1995",]*0.2
 a <- place_x_in_y(x, pop)

}
\seealso{
\code{\link{add_dimension}},\code{\link{add_columns}},\code{\link{mbind}}
}
\author{
Benjamin Bodirsky
}
