% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_magpie.R
\name{population_magpie}
\alias{population_magpie}
\title{population_magpie}
\value{
A2 and B1 population scenario from SRES
}
\description{
Example dataset for a regional MAgPIE object
}
\author{
Benjamin Bodirsky
}
