% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMagpieNC.R
\name{readMagpieNC}
\alias{readMagpieNC}
\title{readMagpieNC}
\usage{
readMagpieNC(filename)
}
\arguments{
\item{filename}{the name of the NetCDF file to be read}
}
\value{
the newly created magpie object
}
\description{
Create a magpie object from a NetCDF file using the "terra" package.
}
\details{
\code{\link{read.magpie}} can also create a magpie object from a NetCDF file, but
it is using the packages ncdf4 and raster, while this function uses terra.
}
\author{
Pascal Sauer
}
