\name{remind2magpie}
\alias{remind2magpie}
\title{Remind2MAgPIE}
\description{Converts a MAgPIE object with Remind regions to a MAgPIE object with MAgPIE regions}
\usage{remind2magpie(x)}
\arguments{
  \item{x}{MAgPIE object with Remind regions}
}
\value{MAgPIE object with MAgPIE regions}
\author{Florian Humpenoeder}
\seealso{\code{"\linkS4class{magpie}"}}
\examples{
 \dontrun{a <- remind2magpie(remind_c_prices)}
}