\name{round-methods}
\docType{methods}
\alias{round-methods}
\alias{round,magpie-method}
\title{ Round-method for MAgPIE objects}
\description{Round-method for MAgPIE-objects respectively. Works exactly as for arrays.
}
\section{Methods}{
\describe{

\item{x = "magpie"}{ works as round(x) for arrays. }

}}

