% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_magpie_units.R
\name{set_magpie_units}
\alias{set_magpie_units}
\title{set_magpie_units (!experimental!)}
\usage{
set_magpie_units(x, value, manual_overwrite = FALSE)
}
\arguments{
\item{x}{MAgPIE object}

\item{value}{object of class units, a character of length one coercible to units via as_units, or a MAgPIE object}

\item{manual_overwrite}{boolean indicating whether to coerce the object into the provided unit. If FALSE (default), 
value must be convertible from x's original unit (or else an error will be thrown), and the data in x will be 
converted to the new unit if possible. If TRUE, value will replace the existing unit without altering the data.}
}
\value{
MAgPIE object x converted to given unit (if possible)
}
\description{
A pipe-friendly version of units<-.magpie. Extension of set_units from the units package to MAgPIE objects.
}
\seealso{
\code{\link[units]{set_units}}
}
\author{
Stephen Bi
}
