% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames_method.R
\docType{methods}
\name{setNames-methods}
\alias{setNames-methods}
\alias{setNames}
\alias{setNames,magpie-method}
\alias{setNames,NULL-method}
\title{Get dataset names}
\usage{
\S4method{setNames}{magpie}(object = nm, nm)
}
\arguments{
\item{object}{MAgPIE object}

\item{nm}{a vector of names current names should be replaced with. If
only one data element exists you can also set the name to NULL.}
}
\description{
Extracts dataset names of a MAgPIE-object
}
\details{
setNames is a shortcut to use a MAgPIE object with manipulated data names.
The setNames method uses the variable names "object" and "nm" in order to be
consistent to the already existing function setNames.
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"ANY\")")}{ normal setNames method }

\item{list("signature(object = \"magpie\")")}{ setNames for MAgPIE objects} }
}

\seealso{
\code{\link{getNames}},
}
\keyword{methods}
