% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCheck.R
\name{sizeCheck}
\alias{sizeCheck}
\title{sizeCheck}
\usage{
sizeCheck(dim, newnames = NULL)
}
\arguments{
\item{dim}{dimensions of the current object as returned by function \code{dim}}

\item{newnames}{a list of new dimensions to be added to the object}
}
\description{
Calculates expected magclass object length and checks that it stays below
the limit defined with magclass_sizeLimit. This is useful to prevent
out of memory errors in case of unwanted object expansions
Ignored if \code{getOption("magclass_sizeLimit")} is negative.
}
\examples{

magclass:::sizeCheck(dim(population_magpie),dimnames(population_magpie))

}
\author{
Jan Philipp Dietrich
}
