% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatRasterToDataset.R
\name{spatRasterToDataset}
\alias{spatRasterToDataset}
\title{spatRasterToDataset}
\usage{
spatRasterToDataset(x)
}
\arguments{
\item{x}{SpatRaster with names of the form "y[0-9]+..[varname]"}
}
\value{
SpatRasterDataset
}
\description{
Convert a SpatRaster to a SpatRasterDataset.
}
\author{
Pascal Sauer
}
