% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressSpecificWarnings.R
\name{suppressSpecificWarnings}
\alias{suppressSpecificWarnings}
\title{suppressSpecificWarnings}
\usage{
suppressSpecificWarnings(expr, regularExpr, fixed = FALSE)
}
\arguments{
\item{expr}{The expression/code to evaluate, can be a block of code inside curly braces.}

\item{regularExpr}{Only warnings matching this regular expression are suppressed.}

\item{fixed}{Match the literal string given by regularExpr instead of interpreting it
as a regular expression. Passed to \code{\link{grepl}}.}
}
\value{
The result of evaluating expr.
}
\description{
Like \code{\link{suppressWarnings}}, but instead of suppressing all warnings this only suppresses
warnings if they match the given pattern.
}
\seealso{
\code{\link{suppressWarnings}}
}
\author{
Pascal Sauer
}
