% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitjoin.R
\name{unitjoin}
\alias{unitjoin}
\title{joins a data.frame or vector of strings with variable and unit separated into
a data.frame with variable and unit joined as 'variable (unit)'.
Use magclass::unitsplit to split them again}
\usage{
unitjoin(x, unit = NULL, col = "variable")
}
\arguments{
\item{x}{data.frame or vector of strings}

\item{unit}{vector of strings. If NULL, col 'unit' in x is used}

\item{col}{column name. Default: variable}
}
\value{
data.frame or vector of strings, dependent on x
}
\description{
joins a data.frame or vector of strings with variable and unit separated into
a data.frame with variable and unit joined as 'variable (unit)'.
Use magclass::unitsplit to split them again
}
