% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.magpie.R
\name{units<-.magpie}
\alias{units<-.magpie}
\alias{units.magpie}
\title{units}
\usage{
\method{units}{magpie}(x) <- value
}
\arguments{
\item{x}{MAgPIE object}

\item{value}{object of class units or character of length one coercible to class units via as_units}
}
\value{
MAgPIE object converted to given unit (if possible)
}
\description{
units method for MAgPIE objects to update the unit of the object
}
\seealso{
\code{\link[units]{units}}
}
\author{
Jan Philipp Dietrich, Stephen Bi
}
