% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitsplit.R
\name{unitsplit}
\alias{unitsplit}
\title{splits a data.frame or vector of strings with form 'variable (unit)' into
a data.frame with variable and unit separated}
\usage{
unitsplit(x, col = "variable")
}
\arguments{
\item{x}{data.frame or vector of strings}

\item{col}{column name. Default: variable}
}
\description{
splits a data.frame or vector of strings with form 'variable (unit)' into
a data.frame with variable and unit separated
}
\author{
Jan Philipp Dietrich, Mika Pflüger, Oliver Richters
}
