% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withAlignedDims.R
\name{withAlignedDims}
\alias{withAlignedDims}
\title{withAlignedDims}
\usage{
withAlignedDims(func, funcName, ...)
}
\arguments{
\item{func}{The function that should be executed with prior alignment of dimensions}

\item{funcName}{The name of the resulting function, to improve error messages}
}
\description{
(for internal use) Executes the passed function after reordering items
in each dimension for all passed magpie objects
}
\author{
Pascal Sauer, Patrick Rein
}
\keyword{internal}
