% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withMetadata.R
\name{withMetadata}
\alias{withMetadata}
\title{withMetadata (!experimental!)}
\usage{
withMetadata(set = NULL, verbosity = NULL)
}
\arguments{
\item{set}{boolean to switch metadata on/off or NULL to leave the option as is.}

\item{verbosity}{Integer to set the verbosity level of calcHistory tracking. 0 = no calcHistory tracking, 
1 = only the core functions are tracked (e.g. calcOutput, readSource), 2 (default) = most magclass functions 
and toolAggregate are also tracked, 3 = virtually all functions are tracked.}
}
\value{
boolean indicating the current metadata setting (switched on or off)
}
\description{
Convenience function to (de-)activate metadata handling in magpie objects and
to return current setting
}
\examples{
\dontrun{
   withMetadata()
   withMetadata(TRUE)
   a <- as.magpie(1)
   getMetadata(a)
   withMetadata(FALSE)
 }
}
\seealso{
\code{\link{getMetadata}}
}
\author{
Jan Philipp Dietrich
}
