% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Wrap}
\usage{
wrap(x, map = list(NA), sep = ".")
}
\arguments{
\item{x}{An array}

\item{map}{A list of length equal to the number of dimensions in the
reshaped array. Each element should be an integer vectors specifying the
dimensions to be joined in corresponding new dimension. One element may
equal NA to indicate that that dimension should be a join of all
non-specified (remaining) dimensions. Default is to wrap everything into a
vector.}

\item{sep}{A character separating joined dimension names}
}
\description{
Reshape an array or a matrix by permuting and/or joining dimensions.
}
\note{
This function is extracted from the R.utils library which is licensed
under LGPL>=2.1 and written by Henrik Bengtsson.
}
\seealso{
\code{\link{unwrap}},\code{\link{fulldim}}

Other MAgPIE-Conversions:
\code{\link[=as.RasterBrick]{as.RasterBrick()}},
\code{\link[=as.SpatRaster]{as.SpatRaster()}},
\code{\link[=as.SpatRasterDataset]{as.SpatRasterDataset()}},
\code{\link[=as.SpatVector]{as.SpatVector()}},
\code{\link{as.array-methods}},
\code{\link{as.data.frame-methods}},
\code{\link[=as_tibble.magpie]{as_tibble.magpie()}},
\code{\link[=unwrap]{unwrap()}}
}
\author{
Henrik Bengtsson, Jan Philipp Dietrich
}
\concept{MAgPIE-Conversions}
