% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.reportProject.R
\name{write.reportProject}
\alias{write.reportProject}
\title{Write file in specific project format}
\usage{
write.reportProject(mif, mapping, file = NULL)
}
\arguments{
\item{mif}{Lists with magpie-objects or a magpie-object as created by read.report or a path to
a report.mif}

\item{mapping}{mapping of the varialbe names of the read in mif. the header
is used for naming.}

\item{file}{name of the project specipic report, default=NULL means that the names of the header of the reporting is used}
}
\description{
Reads in a reporting.mif or uses a magpie object based on a read in
reporting.mif, substitutes names of variables according to the mappping,
mutliplies by an optional factor in the 3rd column of the mapping, and saves
the output in a new *.mif
}
\examples{

\dontrun{
write.reportProject("REMIND_generic_test.mif","Mapping_generic_ADVANCE.csv")
}

}
\seealso{
\code{\link{write.report}}
}
\author{
Christoph Bertram, Lavinia Baumstark, Anastasis Giannousakis
}
