% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMagpieNC.R
\name{writeMagpieNC}
\alias{writeMagpieNC}
\title{writeMagpieNC}
\usage{
writeMagpieNC(
  x,
  filename,
  overwrite = TRUE,
  zname = "time",
  ...,
  compression = 4
)
}
\arguments{
\item{x}{a magpie object}

\item{filename}{the name of the NetCDF file to be created}

\item{overwrite}{whether to overwrite an existing file}

\item{zname}{the name of the time dimension}

\item{...}{additional arguments passed to terra::writeCDF}

\item{compression}{the compression level to be used (1 = no compression, 9 = maximum compression)}
}
\value{
invisibly the SpatRasterDataset created in the process of writing the NetCDF file
}
\description{
Write a magpie object to a NetCDF file using the "terra" package.
}
\details{
\code{\link{write.magpie}} can also write a magpie object to a NetCDF file, but
it is using the packages ncdf4 and raster, while this function uses terra.
}
\author{
Pascal Sauer
#export
}
