% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNC.R
\name{writeNC}
\alias{writeNC}
\title{Write a magpie object to a netCDF file}
\usage{
writeNC(
  x,
  filename,
  unit,
  ...,
  compression = 2,
  missval = NA,
  res = NULL,
  zname = "time"
)
}
\arguments{
\item{x}{A magpie object}

\item{filename}{Name of the netCDF file to write}

\item{unit}{Unit of the data, to omit pass "" (empty string)}

\item{...}{For future expansion.}

\item{compression}{Level of compression to use (1-9), NA for no compression}

\item{missval}{The value that encodes NA in the resulting netCDF file}

\item{res}{Resolution of the data, if not provided it will be guessed}

\item{zname}{Name of the z dimension in the netCDF file}
}
\description{
Write a magpie object to a netCDF file
}
\author{
Pascal Sauer
}
