% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns.R
\name{add_columns}
\alias{add_columns}
\title{add_columns}
\usage{
add_columns(x, addnm = c("new"), dim = 3.1)
}
\arguments{
\item{x}{MAgPIE object which should be extended.}

\item{addnm}{The new columns within dimension "dim"}

\item{dim}{The number of the dimension that should be extended}
}
\value{
The extended MAgPIE object
}
\description{
Function adds new columns to the existing magpie object. The new columns are
filled with NAs.
}
\examples{

 data(population_magpie)
 a <- add_columns(population_magpie)
 str(a)
 fulldim(a)

}
\seealso{
\code{\link{add_dimension}},\code{\link{mbind}}
}
\author{
Benjamin Bodirsky
}
