% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dimension.R
\name{add_dimension}
\alias{add_dimension}
\title{add_dimension}
\usage{
add_dimension(x, dim = 3.1, add = NULL, nm = "dummy", expand = TRUE)
}
\arguments{
\item{x}{MAgPIE object which should be extended.}

\item{dim}{The dimension number of the new dimension (e.g. 3.1)}

\item{add}{The name of the new dimension}

\item{nm}{One or more names of items in the new dimension. If more than one
is given, behavior depends on the expand argument.}

\item{expand}{If TRUE, each item in nm is added to each item
already present, resulting in e.g. `c("A.d1", "B.d1", "A.d2", "B.d2")`.
Otherwise, length of nm must equal the number of items already present and
they are simply added, resulting in e.g. `c("A.d1", "B.d2")`.}
}
\value{
The extended MAgPIE object
}
\description{
Function adds a name dimension as dimension number "dim" with the name "add"
with an empty data column with the name "nm".
}
\examples{

a <- maxample("animal")
str(add_dimension(a, dim = 3.2))
str(add_dimension(a, dim = 2.3, nm = paste0("d", 1:3)))
}
\seealso{
\code{\link{add_columns}},\code{\link{mbind}}
}
\author{
Jan Philipp Dietrich, Benjamin Bodirsky, Pascal Sauer
}
