% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dimension.R
\name{add_dimension}
\alias{add_dimension}
\title{add_dimension}
\usage{
add_dimension(x, dim = 3.1, add = "new", nm = "dummy")
}
\arguments{
\item{x}{MAgPIE object which should be extended.}

\item{dim}{The dimension number of the new dimension. 4 stands for the
second name dimension.}

\item{add}{The name of the new dimension}

\item{nm}{The name of the first entry in dimension "add".}
}
\value{
The extended MAgPIE object
}
\description{
Function adds a name dimension as dimension number "dim" with the name "add"
with an empty data column with the name "nm".
}
\examples{

 a <- add_dimension(maxample("pop"))
 str(a)
 fulldim(a)

}
\seealso{
\code{\link{add_columns}},\code{\link{mbind}}
}
\author{
Benjamin Bodirsky
}
