% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dimension.R
\name{add_dimension}
\alias{add_dimension}
\title{add_dimension}
\usage{
add_dimension(x, dim = 3.1, add = NULL, nm = "dummy")
}
\arguments{
\item{x}{MAgPIE object which should be extended.}

\item{dim}{The dimension number of the new dimension (e.g. 3.1)}

\item{add}{The name of the new dimension}

\item{nm}{The name of the first entry in dimension "add".}
}
\value{
The extended MAgPIE object
}
\description{
Function adds a name dimension as dimension number "dim" with the name "add"
with an empty data column with the name "nm".
}
\examples{

a <- maxample("animal")
str(add_dimension(a, dim = 3.2))
str(add_dimension(a, dim = 2.3, nm = paste0("d", 1:3)))
}
\seealso{
\code{\link{add_columns}},\code{\link{mbind}}
}
\author{
Jan Philipp Dietrich, Benjamin Bodirsky
}
