% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SpatRasterDataset.R
\name{as.SpatRasterDataset}
\alias{as.SpatRasterDataset}
\title{as.SpatRasterDataset}
\usage{
as.SpatRasterDataset(...)
}
\arguments{
\item{...}{arguments passed to as.SpatRaster}
}
\value{
A SpatRasterDataset object
}
\description{
Convert magclass object to a SpatRasterDataset object. Requires the terra package.
}
\details{
Calls \code{\link{as.SpatRaster}} and then \code{\link{spatRasterToDataset}}.
}
\author{
Pascal Sauer
}
