% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.magpie.R
\name{cbind.magpie}
\alias{cbind.magpie}
\title{cbind method for MAgPIE objects}
\usage{
\method{cbind}{magpie}(..., deparse.level = 1)
}
\arguments{
\item{...}{see \link[base]{cbind}}

\item{deparse.level}{see \link[base]{cbind}}
}
\description{
cbind method for MAgPIE-objects. Issues a warning that mbind should be used
if all objects are magpie objects and continues afterwards.
}
\author{
Patrick Rein
}
