% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseDim.R
\name{collapseDim}
\alias{collapseDim}
\title{Collapse dataset dimensions}
\usage{
collapseDim(x, dim = NULL, keepdim = NULL)
}
\arguments{
\item{x}{MAgPIE object}

\item{dim}{Dimension for which the items should be returned. Either number or 
name of dimension or a vector of these. See \code{\link{dimCode}} for more details.
CAUTION: If specified this can lead to duplicate entries. 
If set to NULL all single entry subdimensions will be removed.}

\item{keepdim}{If you want to remove the name of particular dimensions except some, 
you can specify the dimension(s) to keep here. See \code{\link{dimCode}} for naming convention.
Note that \code{keepdim} will be ignored if \code{dim} has been specified.}
}
\value{
The provided MAgPIE object with collapsed dimensions
}
\description{
This function will remove names in the data dimension which are the same for
each element (meaning that this data dimension contains exactly one element)
or, if forced, remove any other subdimension. It is a generalized version
of the function \code{\link{collapseNames}}
}
\examples{

 x <- new.magpie(c("GLO.1","GLO.2"),2000,c("bla.a","bla.b"))  
 collapseDim(x)
 collapseDim(x,keepdim=1:2)
 collapseDim(x,dim=1.1)
 collapseDim(x,dim=3.2)

}
\seealso{
\code{\link{getItems}}
\code{"\linkS4class{magpie}"}
}
\author{
Jan Philipp Dietrich
}
