% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.magpie.R
\name{copy.magpie}
\alias{copy.magpie}
\title{Copy MAgPIE-files}
\usage{
copy.magpie(input_file, output_file, round = NULL)
}
\arguments{
\item{input_file}{file, that should be copied}

\item{output_file}{copy destination}

\item{round}{number of digits the values should be rounded. NULL means no rounding}
}
\description{
This function copies MAgPIE-files from one location to another. During the
copying it is also possible to change the file type (e.g. from 'mz' to
'csv')
}
\examples{

# copy.magpie("bla.csv","blub.mz")
}
\seealso{
\code{\link{read.magpie}},\code{\link{write.magpie}}
}
\author{
Jan Philipp Dietrich
}
