% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimExists.R
\name{dimExists}
\alias{dimExists}
\title{dimExists}
\usage{
dimExists(dim, x, sep = ".")
}
\arguments{
\item{dim}{A vector of dimension numbers or dimension names which should be
checked for.}

\item{x}{MAgPIE object in which the dimensions should be searched for.}

\item{sep}{A character separating joined dimension names.}
}
\value{
Boolean indicating whether the dimension exists or not.
}
\description{
This function checks whether a dimension given by name or number exists.
}
\examples{

pop <- maxample("pop")
dimExists(c("t", "scenario", "blablub"), pop)

}
\seealso{
\code{\link{dimCode}}

Other ObjectInfo: 
\code{\link{fulldim}()},
\code{\link{hasCoords}()},
\code{\link{hasSets}()},
\code{\link{is.temporal}()},
\code{\link{isYear}()},
\code{\link{ncells}()},
\code{\link{ndim}()},
\code{\link{sameDims}()}
}
\author{
Jan Philipp Dietrich
}
\concept{ObjectInfo}
