% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimExists.R
\name{dimExists}
\alias{dimExists}
\title{dimExists}
\usage{
dimExists(dim, x, sep = ".")
}
\arguments{
\item{dim}{A vector of dimension numbers or dimension names which should be
checked for}

\item{x}{MAgPIE object in which the dimensions should be searched for.}

\item{sep}{A character separating joined dimension names}
}
\value{
Boolean indicating whether dimension exists or not
}
\description{
Function checks whether a dimension exsist in a
MAgPIE objects
}
\examples{

pop <- maxample("pop")
dimExists(c("t","scenario","blablub"),pop)

}
\seealso{
\code{\link{dimCode}}
}
\author{
Jan Philipp Dietrich
}
