% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimOrder.R
\name{dimOrder}
\alias{dimOrder}
\title{dimOrder}
\usage{
dimOrder(x, perm)
}
\arguments{
\item{x}{magpie object}

\item{perm}{vector with the new order of the 3rd dimension}
}
\value{
magpie object
}
\description{
Changes the order of the 3rd dimension in a magpie object similar to unwrapping and applying the aperm command, but more efficient.
}
\examples{

\dontrun{ 
pop <- maxample("pop")
x <- setNames(pop,c("kj","kej")) * pop
dimOrder(x=x,perm=c(2,1))
}
}
\author{
Benjamin Leon Bodirsky
}
