% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeRegex.R
\name{escapeRegex}
\alias{escapeRegex}
\title{escapeRegex}
\usage{
escapeRegex(x)
}
\arguments{
\item{x}{String or vector of strings that should be escaped.}
}
\value{
The escaped strings.
}
\description{
Escapes all symbols in a string which have a special meaning in regular
expressions.
}
\seealso{
\code{\link[base]{grep}}
}
\author{
Jan Philipp Dietrich
}
