% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\alias{getCoords<-}
\title{Get Coordinates}
\usage{
getCoords(x, xlab = "x", ylab = "y")

getCoords(x, xlab = "x", ylab = "y") <- value
}
\arguments{
\item{x}{MAgPIE object}

\item{xlab}{label of x-dimension}

\item{ylab}{label of y-dimension}

\item{value}{coordinates as two column data.frame the data should be set to 
(first column = x, second column = y).}
}
\value{
coordinates of the MAgPIE-object
}
\description{
Extracts spatial coordinates of a MAgPIE-object
}
\section{Functions}{
\itemize{
\item \code{getCoords<-}: set coordinates
}}

\examples{

if (requireNamespace("raster", quietly = TRUE)) {
   r <- raster::brick(ncols=360,nrows=180, nl=4)
   r[85:89,176:179] <- (1:20 \%*\% t(1:4))
   names(r) <- c("y2000..bla","y2001..bla","y2000..blub","y2001..blub")
   m <- as.magpie(r)
   getCoords(m)
} 

}
\seealso{
\code{\link{as.RasterBrick}}, \code{\link{getItems}}, \code{"\linkS4class{magpie}"}
}
\author{
Jan Philipp Dietrich
}
