% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasCoords.R
\name{hasCoords}
\alias{hasCoords}
\title{Has Coordinates}
\usage{
hasCoords(x, xlab = "x", ylab = "y")
}
\arguments{
\item{x}{MAgPIE object}

\item{xlab}{label of x-dimension}

\item{ylab}{label of y-dimension}
}
\value{
Boolean indicating whether coordinates were found or not
}
\description{
Checks, whether object contains coordinates.
}
\examples{

hasCoords(population_magpie)

if (requireNamespace("raster", quietly = TRUE)) {
   r <- raster::brick(ncols=360,nrows=180, nl=4)
   r[85:89,176:179] <- (1:20 \%*\% t(1:4))
   names(r) <- c("y2000..bla","y2001..bla","y2000..blub","y2001..blub")
   m <- as.magpie(r)
   hasCoords(m)
} 

}
\seealso{
\code{\link{getCoords}}
}
\author{
Jan Philipp Dietrich
}
