% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasCoords.R
\name{hasCoords}
\alias{hasCoords}
\title{Has Coordinates}
\usage{
hasCoords(x, xlab = "x", ylab = "y")
}
\arguments{
\item{x}{MAgPIE object}

\item{xlab}{label of x-dimension}

\item{ylab}{label of y-dimension}
}
\value{
Boolean indicating whether coordinates were found or not
}
\description{
Checks, whether object contains coordinates.
}
\examples{

hasCoords(maxample("pop"))
hasCoords(maxample("animal"))


}
\seealso{
\code{\link{getCoords}}
}
\author{
Jan Philipp Dietrich
}
