% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasSets.R
\name{hasSets}
\alias{hasSets}
\title{Has Sets}
\usage{
hasSets(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
Boolean indicating whether coordinates were found or not
}
\description{
Checks, whether set names have been set
}
\examples{

hasSets(maxample("pop"))
hasSets(maxample("animal"))

}
\seealso{
\code{\link{getCoords}}

Other ObjectInfo: 
\code{\link{dimExists}()},
\code{\link{fulldim}()},
\code{\link{hasCoords}()},
\code{\link{is.temporal}()},
\code{\link{isYear}()},
\code{\link{ncells}()},
\code{\link{ndim}()},
\code{\link{sameDims}()}
}
\author{
Jan Philipp Dietrich
}
\concept{ObjectInfo}
