% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasSets.R
\name{hasSets}
\alias{hasSets}
\title{Has Sets}
\usage{
hasSets(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
Boolean indicating whether coordinates were found or not
}
\description{
Checks, whether set names have been set
}
\examples{

hasSets(maxample("pop"))
hasSets(maxample("animal"))


}
\seealso{
\code{\link{getCoords}}
}
\author{
Jan Philipp Dietrich
}
