% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.temporal.R
\name{is.temporal}
\alias{is.temporal}
\alias{is.spatial}
\title{is.temporal, is.spatial}
\usage{
is.temporal(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Returns TRUE or FALSE
}
\description{
Functions to find out whether a vector consists of strings consistent with
the definition for auto-detection of temporal or spatial data.
}
\examples{

is.temporal(1991:1993)
is.spatial(c("GLO","AFR"))

}
\author{
Jan Philipp Dietrich
}
