\name{magclass-package}
\alias{magclass-package}
\alias{magclass}
\docType{package}
\title{MAgPIE Class Functions}
\description{Package containing the MAgPIE-Object-Class together with relevant functions and methods.}
\details{
\tabular{ll}{
Package: \tab magclass\cr
Type: \tab Package\cr
Version: \tab 3.74\cr
Date: \tab 2015-11-17\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jan Philipp Dietrich, Benjamin Bodirsky, Markus Bonsch, Henrik Bengtsson, Misko Stevanovic, Florian Humpenoeder

Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
