% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magclass-package.R
\docType{package}
\name{magclass-package}
\alias{magclass-package}
\alias{magclass}
\title{MAgPIE Class Functions}
\description{
Package containing the MAgPIE-Object-Class together with relevant functions
and methods.
}
\details{
\tabular{ll}{ Package: \tab magclass\cr Type: \tab Package\cr Version: \tab
3.93\cr Date: \tab 2016-09-08\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Jan Philipp Dietrich, Benjamin Bodirsky, Misko Stevanovic, Lavinia
Baumstark, Christoph Bertram, Markus Bonsch, Anastasis Giannousakis, Florian
Humpenoeder, David Klein, Ina Neher, Michaja Pehl, Anselm Schultes

Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
