\name{magpieResolution}
\alias{magpieResolution}
\title{magpieResolution}
\description{Returns the Resolution of a MAgPIE object}
\usage{magpieResolution(object)}
\arguments{
  \item{object}{An MAgPIE object}
}
\value{"glo", "reg" or "cell"}
\author{Benjamin Bodirsky}
\seealso{\code{\link{population_magpie}}}
\examples{
data(population_magpie)
magpieResolution(population_magpie)
}

  
