% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpieResolution.R
\name{magpieResolution}
\alias{magpieResolution}
\title{magpieResolution}
\usage{
magpieResolution(object)
}
\arguments{
\item{object}{An MAgPIE object}
}
\value{
"glo", "reg" or "cell"
}
\description{
Returns the Resolution of a MAgPIE object
}
\examples{

magpieResolution(maxample("pop"))

}
\seealso{
\code{\link{maxample}}
}
\author{
Benjamin Bodirsky
}
