% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxample.R
\name{maxample}
\alias{maxample}
\title{maxample}
\usage{
maxample(data)
}
\arguments{
\item{data}{name of the example data set. Currently available are "pop" (regional population data,
previously named "population_magpie"), "animal" (fictional, high-dimensional animal sighting data set)
and "bilateral" (fictional, bilateral trade cost data set).}
}
\value{
the chosen example data set
}
\description{
A collection of magclass example data sets
}
\examples{

p <- maxample("pop")
str(p)

a <- maxample("animal")
str(a)
getItems(a, split = TRUE)
}
\seealso{
Other Display:
\code{\link[=head.magpie]{head.magpie()}},
\code{\link[=mplot]{mplot()}},
\code{\link[=mplotMap]{mplotMap()}},
\code{\link[=print.magpie]{print.magpie()}},
\code{\link{show-methods}},
\code{\link[=str.magpie]{str.magpie()}}
}
\author{
Jan Philipp Dietrich
}
\concept{Display}
