% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{mplot}
\usage{
mplot(px, global = TRUE, total = FALSE)
}
\arguments{
\item{px}{The magpie object to be visualized}

\item{global}{Whether data should be aggregated over regions to global values.}

\item{total}{Whether the total of all data values should also be visualized.}
}
\description{
Get a quick visualization of the content of most magpie objects.
}
\author{
Pascal Sauer, Patrick Rein
}
