% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mPlotMap.R
\name{mPlotMap}
\alias{mPlotMap}
\title{mPlotMap}
\usage{
mPlotMap(px)
}
\arguments{
\item{px}{The magpie object to be visualized. It must contain spatial
coordinates (see \code{\link{hasCoords}}).}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Render a simple world map of a coordinate-based (cell) magpie object. Each
grid cell is drawn as a colored tile on top of country outlines. If the object
contains more than one year or data name, the remaining dimensions are spread
across facets, so it is usually best to subset the object to the slice(s) of
interest before plotting.
}
\examples{
\dontrun{
a <- maxample("animal")
mPlotMap(a[, 1, 1])
}
}
\seealso{
\code{\link{mplot}}, \code{\link{hasCoords}}

Other Display:
\code{\link[=head.magpie]{head.magpie()}},
\code{\link[=maxample]{maxample()}},
\code{\link[=mplot]{mplot()}},
\code{\link[=print.magpie]{print.magpie()}},
\code{\link{show-methods}},
\code{\link[=str.magpie]{str.magpie()}}
}
\author{
Patrick Rein
}
\concept{Display}
