% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncells.R, R/ndata.R, R/nregions.R, R/nyears.R
\name{ncells}
\alias{ncells}
\alias{nyears}
\alias{ndata}
\alias{nregions}
\title{Count elements}
\usage{
ncells(x)

ndata(x)

nregions(x)

nyears(x)
}
\arguments{
\item{x}{A MAgPIE-object}
}
\value{
\item{value}{The number of cells/years/datasets/regions of \code{x}}
}
\description{
Functions to count the number of cells/years/datasets/regions of an
MAgPIE-object
}
\section{Functions}{
\itemize{
\item \code{ndata()}: count datasets

\item \code{nregions()}: count regions

\item \code{nyears()}: count years

}}
\examples{

a <- is.magpie(NULL)
ncells(a)
nyears(a)
ndata(a)
nregions(a)
}
\author{
Jan Philipp Dietrich
}
