% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.magpie.R
\name{print.magpie}
\alias{print.magpie}
\title{print}
\usage{
\method{print}{magpie}(x, drop = TRUE, reshape = FALSE, ...)
}
\arguments{
\item{x}{MAgPIE object}

\item{drop}{argument which controls whether empty dimensions should be
skipped or not.}

\item{reshape}{argument that controls tabular representation of nested data
dimension cross tables, FALSE will reproduce standard print behavior
any pair of two dimension numbers will create a table for these two dims,
and loop over the other dimensions}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Invisibly, the MAgPIE object x.
}
\description{
print method for MAgPIE objects for conventient display of magpie data.
}
\examples{

pop <- maxample("pop")
print(pop)
print(pop[, 1, ], drop = FALSE)
print(pop[, 1, ])
}
\seealso{
\code{\link[base]{print}}
}
\author{
Jan Philipp Dietrich, Kristine Karstens, Felicitas Beier
}
