% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind2magpie.R
\name{remind2magpie}
\alias{remind2magpie}
\title{Remind2MAgPIE}
\usage{
remind2magpie(x)
}
\arguments{
\item{x}{MAgPIE object with Remind regions}
}
\value{
MAgPIE object with MAgPIE regions
}
\description{
Converts a MAgPIE object with Remind regions to a MAgPIE object with MAgPIE
regions
}
\examples{

 \dontrun{a <- remind2magpie(remind_c_prices)}

}
\seealso{
\code{"\linkS4class{magpie}"}
}
\author{
Florian Humpenoeder
}
