% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-method.R
\docType{methods}
\name{round-methods}
\alias{round-methods}
\alias{round,magpie-method}
\title{Round-method for MAgPIE objects}
\usage{
\S4method{round}{magpie}(x, digits = 0)
}
\arguments{
\item{x}{a magpie object}

\item{digits}{integer indicating the number of decimal places (round) or significant
digits (signif) to be used. Negative values are allowed.}
}
\description{
Round-method for MAgPIE-objects respectively. Works exactly as for arrays.
}
\section{Methods}{
 \describe{

\item{x = "magpie"}{ works as round(x) for arrays. }

}
}

