% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameDims.R
\name{sameDims}
\alias{sameDims}
\title{sameDims}
\usage{
sameDims(...)
}
\arguments{
\item{...}{Various magpie objects}
}
\value{
A logical indicating whether the magpie objects have the same items in every dimension.
}
\description{
Checks whether magpie objects have the same items in every dimension (ignoring order).
}
\seealso{
Other ObjectInfo: 
\code{\link{dimExists}()},
\code{\link{fulldim}()},
\code{\link{hasCoords}()},
\code{\link{hasSets}()},
\code{\link{is.temporal}()},
\code{\link{isYear}()},
\code{\link{ncells}()},
\code{\link{ndim}()}
}
\author{
Pascal Sauer, Patrick Rein
}
\concept{ObjectInfo}
