% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameDims.R
\name{sameDims}
\alias{sameDims}
\title{sameDims}
\usage{
sameDims(...)
}
\arguments{
\item{...}{Various magpie objects}
}
\value{
A logical indicating whether the magpie objects have the same items in every dimension.
}
\description{
Checks whether magpie objects have the same items in every dimension (ignoring order).
}
\author{
Pascal Sauer, Patrick Rein
}
