% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show}
\alias{show,ANY-method}
\alias{show,magpie-method}
\title{~~ Method for function \code{show} ~~}
\usage{
\S4method{show}{magpie}(object)
}
\arguments{
\item{object}{A MAgPIE-object}
}
\description{
Show a magpie object by calling the default show method, print, or str,
depending on options("magclass_show_func").
}
\seealso{
Other Display:
\code{\link[=head.magpie]{head.magpie()}},
\code{\link[=maxample]{maxample()}},
\code{\link[=mplot]{mplot()}},
\code{\link[=mplotMap]{mplotMap()}},
\code{\link[=print.magpie]{print.magpie()}},
\code{\link[=str.magpie]{str.magpie()}}
}
\author{
Pascal Sauer
}
\concept{Display}
\keyword{methods}
