% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCheck.R
\name{sizeCheck}
\alias{sizeCheck}
\title{sizeCheck}
\usage{
sizeCheck(dim)
}
\arguments{
\item{dim}{dimensions of the current object as returned by function \code{dim}}
}
\description{
Calculates expected magclass object length and checks that it stays below
the limit defined with magclass_sizeLimit (default = 10^9). This is useful
to prevent out of memory errors in case of unwanted object expansions
Ignored if \code{getOption("magclass_sizeLimit")} is negative.
}
\examples{
pop <- maxample("pop")
magclass:::sizeCheck(dim(pop))

\dontrun{
magclass:::sizeCheck(c(6765L, 10946L, 17711L))
}
}
\author{
Jan Philipp Dietrich
}
