% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.magpie.R
\name{str.magpie}
\alias{str.magpie}
\title{str}
\usage{
\method{str}{magpie}(object, ...)
}
\arguments{
\item{object}{MAgPIE object}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
str method for MAgPIE objects for conventient display of the structure of a magpie object.
}
\details{
In contrast to the default str this will not show the attributes of object@.Data as these
contain only a duplicate of dimnames. Also, dim is not shown, because the information it
provides is implicitly included in dimnames.
}
\examples{
str(maxample("pop"))
}
\seealso{
\code{\link[utils]{str}}
}
\author{
Pascal Sauer
}
