% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap}
\usage{
unwrap(x, sep = ".")
}
\arguments{
\item{x}{A MAgPIE object}

\item{sep}{A character separating joined dimension names}
}
\value{
An array with the full dimensionality of the original data
}
\description{
Reconstruct the full dimensionality of a MAgPIE object
}
\examples{

  a <- as.magpie(array(1:6,c(3,2),list(c("bla","blub","ble"),c("up","down"))))
  fulldim(a)
  unwrap(a)

}
\seealso{
\code{\link{wrap}},\code{\link{fulldim}}
}
\author{
Jan Philipp Dietrich
}
