% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{where}
\usage{
where(x, plot = NULL)
}
\arguments{
\item{x}{A logical statement with a magpie object}

\item{plot}{depreciated. Use the function whereplot in package luplot.}
}
\value{
A list of analysis parameters
}
\description{
Analysis function for magpie objects
}
\examples{

data(population_magpie)
 test<-population_magpie
 dimnames(test)[[1]]<-c("AFG","DEU","FRA","EGY","IND","IDN","RUS","CHN","USA","YEM")
 where(test>500)

}
\seealso{
whereplot in package luplot
}
\author{
Benjamin Leon Bodirsky
}
