% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.report2.R
\name{write.report2}
\alias{write.report2}
\title{Write file in report format}
\usage{
write.report2(...)
}
\arguments{
\item{...}{arguments are forwarded to \code{\link{write.report}}}
}
\description{
This function is deprecated, please use \code{\link{write.report}} instead.
}
\seealso{
\code{\link{write.report}}

Other FileIO: 
\code{\link{copy.magpie}()},
\code{\link{read.magpie}()},
\code{\link{read.report}()},
\code{\link{write.magpie}()},
\code{\link{write.report}()}
}
\author{
Jan Philipp Dietrich
}
\concept{FileIO}
