% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMagpieNC.R
\name{writeMagpieNC}
\alias{writeMagpieNC}
\title{writeMagpieNC}
\usage{
writeMagpieNC(x, filename, overwrite = TRUE, ..., compression = 4)
}
\arguments{
\item{x}{a magpie object}

\item{filename}{the name of the NetCDF file to be created}

\item{...}{additional arguments passed to terra::writeCDF}
}
\value{
invisibly the SpatRasterDataset created in the process of writing the NetCDF file
}
\description{
Write a magpie object to a NetCDF file using the "terra" package.
}
\details{
\code{\link{write.magpie}} can also write a magpie object to a NetCDF file, but
it is using the packages ncdf4 and raster, while this function uses terra.
}
\author{
Pascal Sauer
}
